Graceful plugin and acServer shutdown.
Refactor the plugin kill call to try a SIGINT for some time then use
SIGKILL as required.  In all cases a cmd.Wait is called which ensures
the process is reaped (e.g no defunct procs).

This change mitigates a plugin restart race conditions where the process
being killed is not fully shutdown before a new one is started.  Ie. Issues
with TCP port bind() calls failing when an existing process was not fully terminated.

Related to #1030